face_verify/requirements.txt
fastapi 
uvicorn 
pillow 
numpy 
pandas 
scikit-learn 
face_recognition 
deepface 
python-multipart 
requests 
jinja2
